refactor: Move PartieDeChasseBuilder to its own file
